Rename mlserver.rest.docs to mlserver.rest.openapi
